+++ src/routers/accountsRouter.ts
@@ -942,14 +942,14 @@ accountsRouter.patch('/details/updateDisplayName', async (req: Request, res: Res
   };
 
   if (!userValidation.isValidPassword(requestData.password)) {
-    res.status(400).json({ message: 'Invalid password.' });
+    res.status(400).json({ message: 'Invalid password.', reason: 'password' });
     return;
   };
 
   let connection;
 
   if (!userValidation.isValidDisplayName(requestData.newDisplayName)) {
-    res.status(400).json({ message: 'Invalid display name.' });
+    res.status(400).json({ message: 'Invalid display name.', reason: 'displayName' });
     return;
   };
 
@@ -1024,7 +1024,7 @@ accountsRouter.patch('/details/updateDisplayName', async (req: Request, res: Res
     };
 
     if (requestData.newDisplayName === accountDetails.display_name) {
-      res.status(409).json({ message: `New display name can't be identical to current display name` });
+      res.status(409).json({ message: `Your display name is already ${requestData.newDisplayName}.` });
       return;
     };
 
@@ -1138,12 +1138,12 @@ accountsRouter.patch('/details/updatePassword', async (req: Request, res: Respon
   };
 
   if (!userValidation.isValidPassword(requestData.currentPassword)) {
-    res.status(400).json({ message: 'Invalid password.' });
+    res.status(400).json({ message: 'Invalid password.', reason: 'currentPassword' });
     return;
   };
 
   if (!userValidation.isValidNewPassword(requestData.newPassword)) {
-    res.status(400).json({ message: 'Invalid new password.' });
+    res.status(400).json({ message: 'Invalid new password.', reason: 'newPassword' });
     return;
   };
 
@@ -1160,7 +1160,7 @@ accountsRouter.patch('/details/updatePassword', async (req: Request, res: Respon
         user_type,
         expiry_timestamp
       FROM
-        accounts
+        auth_sessions
       WHERE
         session_id = ?;`,
       [authSessionId]
@@ -1219,16 +1219,12 @@ accountsRouter.patch('/details/updatePassword', async (req: Request, res: Respon
 
     const areIdenticalPasswords: boolean = await bcrypt.compare(requestData.newPassword, accountDetails.hashed_password);
     if (areIdenticalPasswords) {
-      res.status(409).json({
-        message: `New password can't be identical to current password.`,
-        reason: 'identicalPasswords'
-      });
-
+      res.status(409).json({ message: `New password can't be identical to your current password.`, reason: 'identicalPasswords' });
       return;
     };
 
     if (accountDetails.username === requestData.newPassword) {
-      res.status(409).json({ message: `New password can't be identical to username.`, reason: 'passwordEqualsUsername' });
+      res.status(409).json({ message: `New password can't be identical to your username.`, reason: 'passwordEqualsUsername' });
       return;
     };